requirements.txt
altair==4.2.2
numpy==1.24.1
pandas==2.0.1
plotly==5.13.0
psycopg2==2.9.7
psycopg2_binary==2.9.7
pybaseball==2.2.5
pymongo==4.5.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
streamlit==1.18.1
streamlit_authenticator==0.2.3
